feat(modal): activity/project modal are derived from the base Modal component.
The purpose of this is to generalize the modal component as a modal is
only a popup. What's in the pop-up is dependent on the needs of the
popup.
