+++ src/routes/+page.svelte
@@ -1,48 +1,11 @@
 <script lang="ts">
-    import Modal from "$lib/components/Modal.svelte";
     import Timer from "$lib/components/Timer.svelte";
 
     import { clearItems, listAllItems } from "$lib/database/dev";
     import { db } from "$lib/database/db";
     import { seconds } from "$lib/timer";
-
-    let options: string[] = $state([]);
-    let activities = getActivities();
-
-    activities.subscribe({
-        next(activities) {
-            activities.map((activity) => {
-                if (!options.includes(activity.name, 0)) {
-                    options.push(activity.name);
-                }
-            });
-        },
-        error(error) {
-            console.log(`Failed to get activities due to :\n${error}`);
-        },
-        complete() {
-            console.log("Done");
-        },
-    });
-
-    let modalInput: string = $state("");
-    let modal: Modal | null = $state(null);
-    let filteredOptions: string[] = $derived.by(() => {
-        if (modalInput.length === 0) {
-            return options;
-        } else {
-            return options.filter((option) =>
-                option.toLowerCase().includes(modalInput.toLowerCase()),
-            );
-        }
-    });
-
-    let selectedOption: string = $state("Activity");
-
-    function selectOption(option: string) {
-        selectedOption = option;
-        filteredOptions = options;
-    }
+    import ActivityModal from "$lib/components/derived/ActivityModal.svelte";
+    import ProjectModal from "$lib/components/derived/ProjectModal.svelte";
 </script>
 
 <button onclick={async () => await listAllItems(db.activities)}
@@ -51,7 +14,6 @@
 <button
     onclick={async () => {
         await clearItems();
-        options = [];
     }}>Clear db</button
 >
 
@@ -64,56 +26,8 @@
 
     <form class="pt-4">
         <div class="mb-3">
-            <button
-                id="activity select"
-                type="button"
-                class="btn btn-outline-primary w-100"
-                onclick={modal?.showModal}
-            >
-                {selectedOption}
-            </button>
-            <Modal bind:this={modal} title="Select an option">
-                <input
-                    type="text"
-                    class="form-control mb-3"
-                    placeholder="Type to search..."
-                    bind:value={modalInput}
-                />
-                {#if filteredOptions.length > 0 || modalInput.length === 0}
-                    <ul class="options-list">
-                        {#each filteredOptions as option (option)}
-                            <li class="option-item">
-                                <button
-                                    type="button"
-                                    class="btn btn-outline-primary w-100"
-                                    onclick={() => {
-                                        selectOption(option);
-                                        modal?.closeModal();
-                                    }}>{option}</button
-                                >
-                            </li>
-                        {/each}
-                    </ul>
-                {:else}
-                    <div>
-                        <p>No options found for "{modalInput}"</p>
-                        <button
-                            type="button"
-                            class="btn btn-outline-success w-100"
-                            onclick={async () => {
-                                await addActivity(modalInput);
-                                selectOption(modalInput);
-                                modalInput = "";
-                            }}>Create "{modalInput}"</button
-                        >
-                    </div>
-                {/if}
-                <button
-                    type="button"
-                    class="btn btn-secondary w-100 mt-3"
-                    onclick={modal?.closeModal}>Cancel</button
-                >
-            </Modal>
+            <ActivityModal />
+            <ProjectModal />
 
             <div class="goal-container">
                 <span class="goal-label">Goal</span>
@@ -161,15 +75,6 @@
         background-color: green !important;
         color: white !important;
         border-color: green !important;
-    }
-
-    .options-list {
-        list-style: none;
-        padding: 0;
-    }
-
-    .option-item {
-        margin-bottom: 0.5rem;
     }
 
     .goal-container {
